requirements.txt
# Face Emotion Recognition Demo - Requirements

# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.35.0
accelerate>=0.25.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Data Science and Analytics
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Web Application Framework
streamlit>=1.32.0
streamlit-webrtc>=0.47.0

# Plotting and Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# AI Integration
google-genai>=0.3.0
google-generativeai>=0.3.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0

# Performance optimization
psutil>=5.9.0